# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/notifications": {
      "get": {
        "description": "Retrieve paginated notifications for the authenticated user with cursor-based pagination",
        "operationId": "NotificationController_getNotifications",
        "parameters": [
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "description": "Cursor for pagination (notification ID to start from)",
            "schema": {
              "example": "notif123",
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Number of notifications to retrieve (max 50)",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "example": 20,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated notifications retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedNotificationsResponseDto"
                },
                "example": {
                  "data": [
                    {
                      "id": "notif123",
                      "chatId": "chat123",
                      "actorId": "user456",
                      "receiverId": "user123",
                      "type": "MESSAGE",
                      "data": {
                        "messageId": "msg789"
                      },
                      "createdAt": "2023-01-01T12:00:00.000Z",
                      "chat": {
                        "id": "chat123",
                        "title": "Team Chat"
                      },
                      "actor": {
                        "id": "user456",
                        "username": "john_doe"
                      }
                    }
                  ],
                  "meta": {
                    "nextCursor": "notif124",
                    "hasMore": true
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request - Invalid pagination parameters"
          },
          "401": {
            "description": "Unauthorized - Invalid or missing token"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get user notifications",
        "tags": [
          "notifications"
        ]
      }
    },
    "/presence/friends": {
      "get": {
        "description": "Returns a list of online friends for the authenticated user",
        "operationId": "PresenceController_getOnlineFriends",
        "parameters": [],
        "responses": {
          "200": {
            "description": "List of online friend IDs",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Invalid or missing token"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get online friends",
        "tags": [
          "presence"
        ]
      }
    },
    "/presence/bulk": {
      "post": {
        "description": "Get presence status for multiple users at once",
        "operationId": "PresenceController_getBulkPresence",
        "parameters": [],
        "requestBody": {
          "required": true,
          "description": "Array of user IDs to check presence for",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BulkPresenceDto"
              },
              "examples": {
                "example1": {
                  "summary": "Check multiple users",
                  "value": {
                    "userIds": [
                      "user1",
                      "user2",
                      "user3"
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Presence status for all requested users",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BulkPresenceResponseDto"
                },
                "example": {
                  "user1": {
                    "online": true,
                    "lastSeen": null
                  },
                  "user2": {
                    "online": false,
                    "lastSeen": "1672531200000"
                  },
                  "user3": {
                    "online": true,
                    "lastSeen": null
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request - Invalid user IDs"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get bulk presence status",
        "tags": [
          "presence"
        ]
      }
    },
    "/presence/{userId}": {
      "get": {
        "description": "Get presence status for a specific user",
        "operationId": "PresenceController_getPresence",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "description": "User ID to check presence for",
            "schema": {
              "example": "user123",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User presence status",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PresenceResponseDto"
                },
                "example": {
                  "userId": "user123",
                  "online": true,
                  "lastSeen": null
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Invalid or missing token"
          },
          "404": {
            "description": "User not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get user presence",
        "tags": [
          "presence"
        ]
      }
    },
...
  },
  "info": {
    "title": "Chat Backend API",
    "description": "API for chat application",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "jwt"
      }
    },
    "schemas": {
...
      "UserInfoDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "User ID",
            "example": "user123"
          },
          "username": {
            "type": "string",
            "description": "Username",
            "example": "john_doe"
          },
          "email": {
            "type": "string",
            "description": "Email address",
            "example": "john@example.com"
          }
        },
        "required": [
          "id",
          "username",
          "email"
        ]
      },
...
      "ChatInfoDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Chat ID",
            "example": "chat123"
          },
          "title": {
            "type": "string",
            "description": "Chat title",
            "example": "Team Chat"
          }
        },
        "required": [
          "id",
          "title"
        ]
      },
      "NotificationDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Notification ID",
            "example": "notif123"
          },
          "chatId": {
            "type": "string",
            "description": "Chat ID where the notification originated",
            "example": "chat123"
          },
          "actorId": {
            "type": "string",
            "description": "ID of user who triggered the notification",
            "example": "user456"
          },
          "receiverId": {
            "type": "string",
            "description": "ID of user receiving the notification",
            "example": "user123"
          },
          "type": {
            "type": "string",
            "description": "Type of notification",
            "enum": [
              "NEW_CHAT",
              "GROUP_ADDED",
              "GROUP_INVITED",
              "MESSAGE_PINNED",
              "TITLE_UPDATED"
            ],
            "example": "NEW_CHAT"
          },
          "data": {
            "type": "object",
            "description": "Additional notification data",
            "example": {
              "messageId": "msg789"
            },
            "additionalProperties": true
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "When the notification was created",
            "example": "2023-01-01T12:00:00.000Z"
          },
          "chat": {
            "description": "Chat information",
            "allOf": [
              {
                "$ref": "#/components/schemas/ChatInfoDto"
              }
            ]
          },
          "actor": {
            "description": "Actor (user who triggered) information",
            "allOf": [
              {
                "$ref": "#/components/schemas/UserInfoDto"
              }
            ]
          }
        },
        "required": [
          "id",
          "chatId",
          "actorId",
          "receiverId",
          "type",
          "data",
          "createdAt",
          "chat",
          "actor"
        ]
      },
      "PaginatedNotificationsResponseDto": {
        "type": "object",
        "properties": {
          "data": {
            "description": "List of notifications",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/NotificationDto"
            }
          },
          "meta": {
            "type": "object",
            "description": "Pagination metadata",
            "example": {
              "nextCursor": "notif124",
              "hasMore": true
            }
          }
        },
        "required": [
          "data",
          "meta"
        ]
      },
      "BulkPresenceDto": {
        "type": "object",
        "properties": {}
      },
      "BulkPresenceResponseDto": {
        "type": "object",
        "properties": {}
      },
      "PresenceResponseDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}